# gen/typescript/api.schemas.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Product query API
 * API for querying products
 * OpenAPI spec version: 1.1.0
 */
export interface ProductListResponse {
  items: ProductResponse[];
  /** Current page number */
  page: number;
  /** Number of items per page */
  size: number;
  /** Total number of items */
  total: number;
}

export interface ProductResponse {
  id: string;
  name: string;
  description?: string;
  price: number;
  quantity: number;
  /** ID of the product's main image */
  imageId?: string;
  /** Small image URL (400px) for thumbnails and listings */
  smallImageUrl?: string;
  /** Large image URL (800px) for product detail pages */
  largeImageUrl?: string;
  /** ID of the product's category */
  categoryId?: string;
  /** List of attribute values for this product */
  attributes?: AttributeValue[];
}

/**
 * Attribute type for rendering
 */
export type AttributeValueType = typeof AttributeValueType[keyof typeof AttributeValueType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeValueType = {
  single: 'single',
  multiple: 'multiple',
  range: 'range',
  boolean: 'boolean',
  text: 'text',
} as const;

/**
 * How attribute is used (variant for selection, specification for display)
 */
export type AttributeValueRole = typeof AttributeValueRole[keyof typeof AttributeValueRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeValueRole = {
  variant: 'variant',
  specification: 'specification',
} as const;

/**
 * Product attribute with value and display information
 */
export interface AttributeValue {
  /** Reference to the attribute definition */
  attributeId: string;
  /** URL-friendly identifier for filtering */
  slug: string;
  /** Display name of the attribute */
  name: string;
  /** Attribute type for rendering */
  type: AttributeValueType;
  /** Unit for range type attributes */
  unit?: string;
  /** How attribute is used (variant for selection, specification for display) */
  role: AttributeValueRole;
  /** Display order of the attribute in category */
  sortOrder: number;
  /** Attribute values in unified format.
- single: [{slug: "red", value: "Red", colorCode: "#FF0000"}]
- multiple: [{slug: "wifi", value: "WiFi"}, {slug: "bt", value: "Bluetooth"}]
- range: [{value: "128"}]
- boolean: [{value: "true"}] or [{value: "false"}]
- text: [{value: "Custom engraving"}]
 */
  values: AttributeValueItem[];
}

/**
 * Single attribute value
 */
export interface AttributeValueItem {
  /** Option slug for single/multiple types (used for filtering) */
  slug?: string;
  /** Display value (option name, number, boolean, or text) */
  value: string;
  /** Hex color code for color attributes */
  colorCode?: string;
}

export type ProblemErrorsItem = {
  field?: string;
  message?: string;
};

/**
 * RFC7807 Problem Details
 */
export interface Problem {
  type: string;
  title: string;
  status: number;
  detail?: string;
  instance?: string;
  traceId?: string;
  errors?: ProblemErrorsItem[];
}

export interface FacetsResponse {
  /** Available attribute facets computed from actual products */
  facets: AttributeFacet[];
  priceRange: PriceRange;
}

/**
 * Facet for a single attribute showing available values and counts.
Use attribute slug to join with CategoryAttribute from CategoryResponse for display data (name, type, unit, colorCode).

 */
export interface AttributeFacet {
  /** Attribute slug to match with CategoryAttribute.slug */
  slug: string;
  /** Available values with product counts.
- single/multiple: option slug values with counts
- range: numeric values with counts
- boolean: true/false with counts
 */
  values: FacetValue[];
}

/**
 * A single facet value with product count
 */
export interface FacetValue {
  /** Option slug (single/multiple), numeric string (range), or boolean string */
  value: string;
  /** Number of products with this value */
  count: number;
}

/**
 * Price range across all products in the category
 */
export interface PriceRange {
  /** Minimum product price in category */
  min: number;
  /** Maximum product price in category */
  max: number;
}

export type GetRandomProductsParams = {
/**
 * Number of random products to retrieve (1-20)
 * @minimum 1
 * @maximum 20
 */
count: number;
};

export type GetProductListParams = {
/**
 * Page number (must be greater than 0)
 * @minimum 1
 */
page: number;
/**
 * Number of items per page (max 100)
 * @minimum 1
 * @maximum 100
 */
size: number;
/**
 * Filter by category ID
 */
categoryId?: string;
/**
 * Field to sort by
 */
sort?: GetProductListSort;
/**
 * Sort order (ascending or descending)
 */
order?: GetProductListOrder;
/**
 * Minimum price filter
 * @minimum 0
 */
minPrice?: number;
/**
 * Maximum price filter
 * @minimum 0
 */
maxPrice?: number;
/**
 * JSON-encoded array of attribute filters.
Each filter has: slug (string), values (string[]), min (number), max (number).
Example: [{"slug":"color","values":["red","blue"]},{"slug":"storage","min":128,"max":512}]

 */
attributeFilters?: string;
};

export type GetProductListSort = typeof GetProductListSort[keyof typeof GetProductListSort];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetProductListSort = {
  name: 'name',
  price: 'price',
  quantity: 'quantity',
} as const;

export type GetProductListOrder = typeof GetProductListOrder[keyof typeof GetProductListOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetProductListOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type GetProductFacetsParams = {
/**
 * Category ID to compute facets for
 */
categoryId: string;
};